requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-socketio==5.11.4
aiohttp==3.10.5
jinja2==3.1.4
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
bcrypt==4.1.2